feat(PNC-26): DeckView is passed deck names that are read is from the settings
