+++ client/src/components/settings-page.tsx
@@ -8,11 +8,16 @@ import { useEffect, useState } from "react";
 import { Notification } from "./notification";
 import Container from "react-bootstrap/Container";
 
-export function SettingsPage() {
-    const [settings, setSettings] = useState<Settings>();
+interface SettingsPageProps {
+    settings: Settings | undefined,
+    setSettings: React.Dispatch<React.SetStateAction<Settings | undefined>>
+}
+
+export function SettingsPage({ settings, setSettings} : SettingsPageProps) {
+    //const [settings, setSettings] = useState<Settings>();
     const [showNotification, setShowNotification] = useState(false);
     const [notification, setNotification] = useState("");
-
+    /*
     useEffect(() => {
         const getSettingsAsync = async () => {
             const settingsData = await readSettings();
@@ -22,7 +27,7 @@ export function SettingsPage() {
         } 
         getSettingsAsync();
     }, [])
-
+    */
     const writeSettingsAsync = async() => {
         if (settings) {
             const success = await writeSettings(settings);
